python_test/requirements.txt
Django==2.1.8
django-extensions==2.2.1
djangorestframework==3.10.3